+++ src/main/java/backend/Logic.java
@@ -31,6 +31,7 @@ public class Logic {
 
     private RepoIO repoIO;
     public LoginController loginController;
+    public UpdateController updateController;
 
     public Logic(UIManager uiManager, Preferences prefs, boolean isTestMode, boolean enableTestJSON) {
         this.uiManager = uiManager;
@@ -39,6 +40,7 @@ public class Logic {
 
         repoIO = new RepoIO(isTestMode, enableTestJSON);
         loginController = new LoginController(this);
+        updateController = new UpdateController(this);
 
         // Only relevant to testing, need a different event type to avoid race condition
         UI.events.registerEvent((ClearLogicModelEventHandler) e -> {
@@ -171,7 +173,7 @@ public class Logic {
                 forEach(repoIdNotInUse -> models.removeRepoModelById(repoIdNotInUse));
     }
 
-    private ImmutablePair<Integer, Long> updateRemainingRate
+    public ImmutablePair<Integer, Long> updateRemainingRate
             (ImmutablePair<Integer, Long> rateLimits) {
         uiManager.updateRateLimits(rateLimits);
         return rateLimits;
@@ -223,70 +225,18 @@ public class Logic {
      * and then sends the data to the GUI.
      */
     private void refreshUI() {
-        filterSortRefresh(getAllUIFilters());
+        updateController.filterSortRefresh(getAllUIFilters());
     }
 
     /**
-     * Given a list of filter expressions, dispatch metadata update if needed and then process them to return
-     * a map of filtered and sorted issues corresponding to each filter expression, based on the most recent data
-     * from the repository source.
+     * Feeds a one-element list of filter expressions to updateController.
      *
-     * @param filterExprs Filter expressions to process
+     * @param filterExpr The filter expression to be processed by updateController.
      */
-    public void filterSortRefresh(List<FilterExpression> filterExprs) {
-        // Open specified repos
-        openRepositoriesInFilters(filterExprs);
-
-        // First filter, for issues requiring a metadata update.
-        Map<String, List<TurboIssue>> toUpdate = tallyMetadataUpdate(filterExprs);
-
-        if (toUpdate.size() > 0) {
-            // If there are issues requiring metadata update, we dispatch the metadata requests...
-            ArrayList<CompletableFuture<Boolean>> metadataRetrievalTasks = new ArrayList<>();
-            toUpdate.forEach((repoId, issues) -> {
-                metadataRetrievalTasks.add(getIssueMetadata(repoId, issues));
-            });
-            // ...and then wait for all of them to complete.
-            Futures.sequence(metadataRetrievalTasks)
-                    .thenAccept(results -> logger.info("Metadata retrieval successful for "
-                            + results.stream().filter(result -> result).count() + "/"
-                            + results.size() + " repos"))
-                    .thenCompose(n -> getRateLimitResetTime())
-                    .thenApply(this::updateRemainingRate)
-                    .thenRun(() -> updateUI(filterAndSort(filterExprs))); // Then filter the second time.
-        } else {
-            // If no issues requiring metadata update, just run the filter and sort.
-            updateUI(filterAndSort(filterExprs));
-        }
-    }
-
-    /**
-     * Given a list of filter expressions, open all repositories necessary for processing the filter expressions.
-     *
-     * @param filterExprs Filter expressions to process.
-     */
-    private void openRepositoriesInFilters(List<FilterExpression> filterExprs) {
-        filterExprs.stream()
-                .flatMap(filterExpr -> Qualifier.getContentOfMetaQualifier(filterExpr, Qualifier.REPO).stream())
-                .distinct()
-                .forEach(this::openRepositoryFromFilter);
-    }
-
-    /**
-     * Given a list of filter expressions, determine issues within the model that require a metadata update.
-     *
-     * @param filterExprs Filter expressions to process for metadata requests.
-     * @return Repo IDs and the corresponding issues in the repo requiring a metadata update.
-     */
-    private Map<String, List<TurboIssue>> tallyMetadataUpdate(List<FilterExpression> filterExprs) {
-        List<TurboIssue> allModelIssues = models.getIssues();
-
-        return filterExprs.stream()
-                .filter(Qualifier::hasUpdatedQualifier)
-                .flatMap(filterExpr -> allModelIssues.stream()
-                        .filter(issue -> Qualifier.process(models, filterExpr, issue)))
-                .distinct()
-                .collect(Collectors.groupingBy(TurboIssue::getRepoId));
+    public void refreshPanel(FilterExpression filterExpr) {
+        List<FilterExpression> panelExpr = new ArrayList<>();
+        panelExpr.add(filterExpr);
+        updateController.filterSortRefresh(panelExpr);
     }
 
     /**
@@ -316,64 +266,9 @@ public class Logic {
     }
 
     /**
-     * Filters and sorts issues within the model according to the given filter expressions.
-     *
-     * @param filterExprs Filter expressions to process.
-     * @return Filter expressions and their corresponding issues after filtering and sorting.
+     * Carries the current model in Logic, as well as issues to be displayed in panels, to the GUI.
      */
-    private Map<FilterExpression, List<TurboIssue>> filterAndSort(List<FilterExpression> filterExprs) {
-        List<TurboIssue> allModelIssues = models.getIssues();
-
-        Map<FilterExpression, List<TurboIssue>> filteredAndSorted = new HashMap<>();
-
-        filterExprs.forEach(filterExpr -> {
-            List<TurboIssue> filterAndSortedExpression = filteredAndSorted.get(filterExpr);
-
-            if (filterAndSortedExpression == null) { // If it already exists, no need to filter anymore
-                boolean hasUpdatedQualifier = Qualifier.hasUpdatedQualifier(filterExpr);
-
-                List<TurboIssue> filteredAndSortedIssues = allModelIssues.stream()
-                        .filter(issue -> Qualifier.process(models, filterExpr, issue))
-                        .sorted(determineComparator(filterExpr, hasUpdatedQualifier))
-                        .collect(Collectors.toList());
-
-                filteredAndSorted.put(filterExpr, filteredAndSortedIssues);
-            }
-        });
-
-        return filteredAndSorted;
-    }
-
-    /**
-     * Produces a suitable comparator based on the given filter expression.
-     *
-     * @param filterExpr The given filter expression.
-     * @param hasUpdatedQualifier Determines the behaviour of the sort key "nonSelfUpdate".
-     * @return The comparator to use.
-     */
-    private Comparator<TurboIssue> determineComparator(FilterExpression filterExpr,
-                                                       boolean hasUpdatedQualifier) {
-        for (Qualifier metaQualifier : filterExpr.find(Qualifier::isMetaQualifier)) {
-            // Only take into account the first sort qualifier found
-            if (metaQualifier.getName().equals("sort")) {
-                return metaQualifier.getCompoundSortComparator(models, hasUpdatedQualifier);
-            }
-        }
-
-        // No sort qualifier, look for updated qualifier
-        if (hasUpdatedQualifier) {
-            return Qualifier.getSortComparator(models, "nonSelfUpdate", true, true);
-        }
-
-        // No sort or updated, return sort by descending ID, which is the default.
-        return Qualifier.getSortComparator(models, "id", true, false);
-    }
-
-    /**
-     * Carries the current model in Logic to the GUI and triggers metadata updates if panels require
-     * metadata to display their issues, in which case the changes in the model are not presented to the user.
-     */
-    private void updateUI(Map<FilterExpression, List<TurboIssue>> issuesToShow) {
+    public void updateUI(Map<FilterExpression, List<TurboIssue>> issuesToShow) {
         uiManager.update(models, issuesToShow);
     }
 
@@ -384,5 +279,14 @@ public class Logic {
      */
     private List<FilterExpression> getAllUIFilters() {
         return uiManager.getAllFilters();
+    }
+
+    /**
+     * For use by UpdateController to perform filtering.
+     *
+     * @return The currently held MultiModel.
+     */
+    public MultiModel getModels() {
+        return models;
     }
 }
